+++ hombreslobo-typescript/index.html
@@ -4,20 +4,25 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Castronegro – Hombres Lobo</title>
-    <link href="https://fonts.googleapis.com/css2?family=Bahiana&display=swap" rel="stylesheet">
-    <link rel="icon" href="src/imagenes/logoLobo.png" type="image/x-icon">
+    <link
+      href="https://fonts.googleapis.com/css2?family=Bahiana&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="icon" href="/src/imagenes/logoLobo.png" type="image/x-icon" />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
     />
-    <link rel="stylesheet" href="./css/base.css" />
-    <link rel="stylesheet" href="./css/index.css" />
   </head>
   <body>
-
     <header class="cabecera">
       <div class="nav-logo-titulo">
-        <a href="index.html"><img src="src/imagenes/logoLobo.png" class="logo" alt="Logo Castronegro"></a>
+        <a href="/index.html"
+          ><img
+            src="/src/imagenes/logoLobo.png"
+            class="logo"
+            alt="Logo Castronegro"
+        /></a>
         <h1 class="titulo-sitio">Castronegro</h1>
       </div>
       <nav class="menu">
@@ -25,7 +30,7 @@
         <a href="#">Reglas</a>
         <a href="#">Comunidad</a>
         <a href="#">Iniciar Sesión</a>
-        <a href="./registro.html">Regístrate</a>
+        <a href="/src/autenticacion/htmls/registro.html">Regístrate</a>
       </nav>
     </header>
 
@@ -36,7 +41,7 @@
           El engaño, la sospecha y la supervivencia te esperan. ¿Tienes lo
           necesario para descubrir al lobo?
         </p>
-        <img src="src/imagenes/cartas-lobos.png" alt="Cartas de Lobos">
+        <img src="/src/imagenes/cartas-lobos.png" alt="Cartas de Lobos" />
         <button class="boton-principal">¡Jugar Ahora!</button>
       </div>
     </section>
@@ -82,6 +87,6 @@
       </div>
     </footer>
 
-    <script src="src/main.ts"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
