+++ hombreslobo-typescript/src/autenticacion/htmls/registro.html
@@ -8,23 +8,29 @@
       href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;700&family=Merriweather:wght@700&display=swap"
       rel="stylesheet"
     />
-    <link href="https://fonts.googleapis.com/css2?family=Bahiana&display=swap" rel="stylesheet">
-    <link rel="icon" href="src/imagenes/logoLobo.png" type="image/x-icon">
+    <link
+      href="https://fonts.googleapis.com/css2?family=Bahiana&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="icon" href="/src/imagenes/logoLobo.png" type="image/x-icon" />
     <link
       rel="stylesheet"
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.0/css/all.min.css"
     />
-    <link rel="stylesheet" href="./css/base.css" />
-    <link rel="stylesheet" href="./css/registro.css" />
   </head>
   <body>
     <header class="cabecera">
       <div class="nav-logo-titulo">
-      <a href="index.html"><img src="src/imagenes/logoLobo.png" class="logo" alt="Logo Castronegro"></a>
+        <a href="/index.html"
+          ><img
+            src="/src/imagenes/logoLobo.png"
+            class="logo"
+            alt="Logo Castronegro"
+        /></a>
         <h1 class="titulo-sitio">Castronegro</h1>
       </div>
       <nav class="menu">
-        <a href="index.html">Inicio</a>
+        <a href="/index.html">Inicio</a>
         <a href="#">Reglas</a>
         <a href="#">Comunidad</a>
         <a href="#">Iniciar Sesión</a>
@@ -108,7 +114,7 @@
             />
           </div>
 
-          <button type="submit" class="boton-principal">Registrarse</button>
+          <button type="submit" class="boton-registrar">Registrarse</button>
           <p class="enlace-login">
             ¿Ya tienes cuenta? <a href="#">Inicia sesión</a>
           </p>
@@ -120,6 +126,6 @@
       <p>© 2025 Hombres Lobo Online. Todos los derechos reservados.</p>
     </footer>
 
-    <script type="module" src="./src/main.ts"></script>
+    <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
