+++ hombreslobo-typescript/src/main.ts
@@ -1,5 +1,7 @@
-// import './style.css'
-import { validarFormulario } from "./modules/validarFormulario";
+import "./css/base.css";
+import "./css/index.css";
+import "/src/autenticacion/css/registro.css";
+import { validarFormulario } from "./autenticacion/validarFormulario";
 
 document.addEventListener("DOMContentLoaded", () => {
   validarFormulario();
